Finish rt-app backend, general refactorings

+++ test_suite_rs/src/tests/generic/runner_args.rs
@@ -52,7 +52,7 @@ pub struct RunnerArgsBase {
     /// the bandwidth is reserved for SCHED_OTHER tasks and the other 5% is used
     /// for overheads (?).
     #[arg(short = 'b', long = "max-bw", value_name = "f32", default_value = "0.90")]
-    pub max_allocable_bw: f32,
+    pub max_allocable_bw: f64,
 
     /// number of instances per job
     #[arg(short = 'j', long = "job", value_name = "u64", default_value = "200")]
